[Spring] show matieres enseignes par enseignant

+++ Spring/src/main/java/com/projetIntegraion/spring/Etudiant/demandeCreationClub/repository/MatiereRepository.java
@@ -12,7 +12,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 public interface MatiereRepository extends JpaRepository<Matiere, Long> {
     Page<Matiere> findByClasseId(Long id,Pageable pageable);
     Boolean existsByNom(String nom);
-    Page<Matiere> findByEnseignantId(Long id,Pageable pageable);
+    Page<Matiere> findByEnseignantId(Integer id,Pageable pageable);
 
     
 }
